feature(reintegrate-gestaltv7): fix package names for modules
