+++ src/main/java/org/terasology/cooperandbronze/system/CopperAndBronzeJournalIntegration.java
@@ -1,6 +1,6 @@
 // Copyright 2020 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
-package org.terasology.cab.system;
+package org.terasology.cooperandbronze.system;
 
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.event.ReceiveEvent;
